examples/gemini_chat/requirements.txt
askdiana[app]
python-dotenv
google-genai
